added SwiftList and SwiftItem
also fixed bags
