+++ messages/en.json
@@ -82,6 +82,10 @@
     "title": "OUR Programs",
     "description": "Our charitable foundation is dedicated to supporting people who find themselves in difficult life circumstances. We provide assistance in various areas, from humanitarian aid to educational and medical initiatives. Our mission is to create conditions for a dignified life for every individual, helping them overcome challenges and discover new opportunities for growth. Through our projects, we aim to improve the quality of life for vulnerable groups, offering hope and support to those who need it the most."
   },
+  "ButtonCopy": {
+    "copy": "Copy",
+    "copied": "Copied"
+  },
   "ErrorMessages": {
     "nothingWasFoundFor": "Nothing was found for"
   }
