+++ messages/ua.json
@@ -82,6 +82,10 @@
     "title": "Напрямки роботи",
     "description": "Наш благодійний фонд спрямований на підтримку людей, які опинилися у складних життєвих обставинах. Ми надаємо допомогу у різних сферах, від гуманітарної підтримки до освітніх та медичних ініціатив. Наша місія — створювати умови для гідного життя кожної людини, допомагаючи долати труднощі та відкривати нові можливості для розвитку. Через наші проєкти ми прагнемо покращити якість життя вразливих категорій населення, даруючи надію та підтримку тим, хто цього потребує найбільше."
   },
+  "ButtonCopy": {
+    "copy": "Копіювати",
+    "copied": "Скопійовано"
+  },
   "ErrorMessages": {
     "nothingWasFoundFor": "Нічого не було знайдено за запитом"
   }
