+++ src/core/SwiftItem/SwiftItem.module.css
@@ -1,0 +1,84 @@
+.listItem {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.listItem:not(:last-child) {
+  margin-bottom: 40px;
+}
+
+.listItem:nth-child(1),
+.listItem:nth-child(3) {
+  margin-bottom: 24px;
+}
+
+.listTitle {
+  font-weight: 500;
+  font-size: 24px;
+  color: var(--blue);
+  margin-bottom: 24px;
+}
+
+.listItem:nth-child(5) div h3,
+.listItem:nth-child(6) div h3,
+.listItem:nth-child(7) div h3 {
+  margin-bottom: 8px;
+}
+
+.subTitle {
+  font-weight: 500;
+  font-size: 24px;
+  color: var(--text);
+  margin-bottom: 8px;
+}
+
+.listText {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+}
+
+.listItem:nth-child(1) .listText {
+  max-width: 331px;
+}
+
+.listItem:nth-child(4) .listText li:nth-child(2),
+.listItem:nth-child(2) .listText li:nth-child(2) {
+  max-width: 193px;
+}
+
+.listItem:nth-child(5) .listText li:nth-child(1) {
+  max-width: 442px;
+}
+
+.listText li {
+  font-weight: 500;
+  font-size: 20px;
+  line-height: 1.5;
+  color: var(--text);
+  max-width: 422px;
+}
+
+.btnCopy {
+  display: flex;
+  gap: 8px;
+  background-color: transparent;
+  height: 24px;
+  border: none;
+  font-weight: 500;
+  font-size: 20px;
+  text-align: center;
+  color: var(--blue);
+  cursor: pointer;
+  transition: all 0.3s ease-out;
+}
+
+.btnCopy:focus,
+.btnCopy:hover {
+  color: var(--card-fill-tap);
+}
+
+.btnCopy:active {
+  color: var(--text);
+}
