+++ src/core/SwiftItem/SwiftItem.tsx
@@ -1,0 +1,65 @@
+import { useTranslations } from "next-intl";
+import { useState } from "react";
+import { FaRegCopy } from "react-icons/fa";
+
+import { montserrat } from "../fonts";
+
+import css from "./SwiftItem.module.css";
+
+type PropsType = {
+  data: {
+    id: string;
+    title: null | string;
+    subTitle: null | string;
+    texts: string[];
+  };
+};
+
+export default function SwiftItem({ data }: PropsType) {
+  const translate = useTranslations("ButtonCopy");
+
+  const [copied, setCopied] = useState(false);
+  const { title, subTitle, texts } = data;
+
+  const copyToClipboard = async (textArr: string[]) => {
+    if (copied) return;
+    const text = textArr.join(" ");
+
+    try {
+      await navigator.clipboard.writeText(text);
+      setCopied(true);
+      setTimeout(() => setCopied(false), 2000);
+    } catch (err) {
+      console.error("Помилка копіювання: ", err);
+    }
+  };
+
+  return (
+    <li className={css.listItem}>
+      <div>
+        {title && (
+          <h3 className={`${montserrat.className} ${css.listTitle}`}>
+            {title}
+          </h3>
+        )}
+        {subTitle && (
+          <p className={`${montserrat.className} ${css.subTitle}`}>
+            {subTitle}
+          </p>
+        )}
+        <ul className={`${montserrat.className} ${css.listText}`}>
+          {texts.map((text, index) => (
+            <li key={index}>{text}</li>
+          ))}
+        </ul>
+      </div>
+      <button
+        className={`${montserrat.className} ${css.btnCopy}`}
+        onClick={() => copyToClipboard(texts)}
+      >
+        <FaRegCopy size="24px" />
+        {copied ? translate("copied") : translate("copy")}
+      </button>
+    </li>
+  );
+}
